+++ docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # What's New with Enterprise-NG
 
+## 18.8.0
+
+## Fixes
+
+- `[Modal]` Added `noPadding` setting to modal. ([#6201](https://inforwiki.atlassian.net/browse/IDS-6201))
+
 ## 18.7.0
 
 ## 18.7.0 Fixes